Create sndinfo.txt
If you ever wanted your chaingun in Doom 1/2 sounds like a chaingunner's - run this txt file when you open GZDoom, which will change the sound to a chaingun (and leave the original sound for pistol). After that, the chaingun sound will be like a shotgun's (the reloading sound at the end is okay - everyone in the game has it). You can run this file using args.