Update sndinfo.txt
Tired of the default chaingun sound? If you ever wanted your chaingun in Doom 1/2 sounds like enemy's - run this txt file when you open GZDoom (or paste the code to the new sndinfo.txt), which will change the chaingun sound. After that, the sound will be like a shotgun's (the reloading sound at the end is okay - everyone in the game has it). The pistol sound will remain default. You can run this file using args.